l03_0_mcp_stdio/requirements.txt
google-adk
python-whois
mcp
